pyproject.toml
[project]
name = "s13p11a102"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rpi-gpio>=0.7.1",
    "rplcd>=1.4.0",
    "smbus>=1.1.post2",
]
